# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PIX Hub API",
        "description": "API para integração com subadquirentes de pagamento (PIX e Saques)",
        "contact": {
            "email": "dev@pixhub.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/pix": {
            "get": {
                "tags": [
                    "PIX"
                ],
                "summary": "Listar transações PIX",
                "description": "Lista todas as transações PIX do usuário autenticado",
                "operationId": "listPix",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número máximo de registros",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de transações PIX",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "external_id": {
                                                        "type": "string"
                                                    },
                                                    "amount": {
                                                        "type": "number"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PIX"
                ],
                "summary": "Criar um novo PIX",
                "description": "Cria uma nova transação PIX e retorna o QR Code para pagamento",
                "operationId": "createPix",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Valor do PIX",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "description": {
                                        "description": "Descrição do pagamento",
                                        "type": "string",
                                        "example": "Pagamento teste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PIX criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PIX criado com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "external_id": {
                                                    "type": "string",
                                                    "example": "PIX_abc123"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 100.5
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "qr_code": {
                                                    "type": "string",
                                                    "example": "00020126580014br.gov.bcb.pix..."
                                                },
                                                "qr_code_base64": {
                                                    "type": "string",
                                                    "example": "data:image/png;base64,..."
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2025-11-19T04:30:00.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2025-11-19T04:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O campo amount é obrigatório."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Muitas requisições (Rate Limit: 60 req/min)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too Many Attempts."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pix/{id}": {
            "get": {
                "tags": [
                    "PIX"
                ],
                "summary": "Obter detalhes de um PIX",
                "description": "Retorna os detalhes de uma transação PIX específica",
                "operationId": "getPix",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da transação PIX",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do PIX",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "external_id": {
                                                    "type": "string",
                                                    "example": "PIX_abc123"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 100.5
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "paid"
                                                },
                                                "qr_code": {
                                                    "type": "string"
                                                },
                                                "paid_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PIX não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "PIX não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Acesso negado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação via Laravel Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}